+++ inshd/src/request_handlers/mod.rs
@@ -1,0 +1,12 @@
+//! Handlers for individual request types.
+mod create_file;
+mod find_files;
+mod get_files;
+mod search_phrase;
+mod suggest_search_phrase;
+
+pub use create_file::CreateFile;
+pub use find_files::FindFiles;
+pub use get_files::GetFiles;
+pub use search_phrase::SearchPhrase;
+pub use suggest_search_phrase::SuggestSearchPhrase;